Move error into design library
